prune usecase 1
