+++ libs/python/btp_cli.py
@@ -4,7 +4,7 @@ from libs.python.helperJson import addKeyValuePair, dictToString, convertStringT
 from libs.python.helperBtpTrust import delete_cf_service_key, runTrustFlow, get_cf_service_key
 from libs.python.helperCommandExecution import executeCommandsFromUsecaseFile, runShellCommand, runCommandAndGetJsonResult, runShellCommandFlex, login_btp, login_cf
 from libs.python.helperEnvCF import checkIfCFEnvironmentAlreadyExists, checkIfCFSpaceAlreadyExists, try_until_cf_space_done, get_cf_service_deletion_status
-from libs.python.helperServiceCreation import initiateCreationOfServiceInstances, checkIfAllServiceInstancesCreated
+from libs.python.helperServiceInstances import initiateCreationOfServiceInstances, checkIfAllServiceInstancesCreated
 from libs.python.helperGeneric import buildUrltoSubaccount, getNamingPatternForServiceSuffix, createSubaccountName, createSubdomainID, createOrgName, getTimingsForStatusRequest, save_collected_metadata
 from libs.python.helperFileAccess import writeKubeConfigFileToDefaultDir
 from libs.python.helperEnvKyma import extractKymaDashboardUrlFromEnvironmentDataEntry, getKymaEnvironmentInfoByClusterName, getKymaEnvironmentStatusFromEnvironmentDataEntry, extractKymaKubeConfigUrlFromEnvironmentDataEntry, getKymaEnvironmentIdByClusterName
@@ -1072,61 +1072,59 @@ def pruneUseCaseAssets(btpUsecase: BTPUSECASE):
         log.info("Delete service instances")
        
         # Initiate deletion of service instances
-        for environment in btpUsecase.definedEnvironments:
-            if environment.name == "cloudfoundry":
-                for service in accountMetadata["createdServiceInstances"]:
-                    if "createdServiceKeys" in service:
-                        for key in service["createdServiceKeys"]:
-                            delete_cf_service_key(btpUsecase, service["instancename"], key["keyname"])
-                        search_every_x_seconds, usecaseTimeout = getTimingsForStatusRequest(btpUsecase, service)
-                        current_time = 0
-                        while usecaseTimeout > current_time:
-                            command = "cf service-key '" + service["instancename"] + "' " + key["keyname"]
-                            # Calling the command with the goal to get back the "FAILED" status, as this means that the service key was not found (because deletion was successfull)
-                            # If the status is not "FAILED", this means that the deletion hasn't been finished so far
-                            message = "check if service key >" + key["keyname"] + "< for service instance >" + service["instancename"] + "<"
-                            p = runShellCommandFlex(btpUsecase, command, "CHECK", message, False, False)
-                            result = p.stdout.decode()
-                            if "FAILED" in result:
-                                usecaseTimeout = current_time - 1
-                            time.sleep(search_every_x_seconds)
-                            current_time += search_every_x_seconds
-                    if "instancename" in service and service["instancename"] is not None and service["instancename"] != "":
-                        command = "cf delete-service '" + service["instancename"] + "' -f"
-                        message = "Delete CF service instance >" + service["instancename"] + "< from subaccount"
-                        result = runShellCommand(btpUsecase, command, "INFO", message)
-
-                log.info("Check deletion status for service instances")
-
-                # check status of deletion
-                search_every_x_seconds = btpUsecase.repeatstatusrequest
-                usecaseTimeout = btpUsecase.repeatstatustimeout
-                current_time = 0
-                allServicesDeleted = False
-                # Set the deletion status to "not deleted"
-                for service in accountMetadata["createdServiceInstances"]:
-                    service["deletionStatus"] = "not deleted"
-
-                while usecaseTimeout > current_time and allServicesDeleted is False:
-                    for service in accountMetadata["createdServiceInstances"]:
-                        if "instancename" not in service:
-                            status = "deleted"
-                            service["deletionStatus"] = status
-                            log.info("no service instance available for service >" + service["name"] + "<. Deletion not needed.")
-                            continue
-                        status = get_cf_service_deletion_status(btpUsecase, service)
-                        if (status == "deleted"):
-                            log.success("service instance >" + service["instancename"] + "< for service >" + service["name"] + "< now deleted.")
-                            service["deletionStatus"] = "deleted"
-                        else:
-                            service["deletionStatus"] = status
-                    time.sleep(search_every_x_seconds)
-                    current_time += search_every_x_seconds
-                    allServicesDeleted = True
-                    for service in accountMetadata["createdServiceInstances"]:
-                        if service["deletionStatus"] != "deleted":
-                            allServicesDeleted = False
-                log.success("all service instances now deleted.")
+        for service in accountMetadata["createdServiceInstances"]:
+            if "createdServiceKeys" in service:
+                for key in service["createdServiceKeys"]:
+                    if service.targetenvironment == "cloudfoundry":    
+                        delete_cf_service_key(btpUsecase, service["instancename"], key["keyname"])
+                    search_every_x_seconds, usecaseTimeout = getTimingsForStatusRequest(btpUsecase, service)
+                    current_time = 0
+                    while usecaseTimeout > current_time:
+                        command = "cf service-key '" + service["instancename"] + "' " + key["keyname"]
+                        # Calling the command with the goal to get back the "FAILED" status, as this means that the service key was not found (because deletion was successfull)
+                        # If the status is not "FAILED", this means that the deletion hasn't been finished so far
+                        message = "check if service key >" + key["keyname"] + "< for service instance >" + service["instancename"] + "<"
+                        p = runShellCommandFlex(btpUsecase, command, "CHECK", message, False, False)
+                        result = p.stdout.decode()
+                        if "FAILED" in result:
+                            usecaseTimeout = current_time - 1
+                        time.sleep(search_every_x_seconds)
+                        current_time += search_every_x_seconds
+            
+            if "instancename" in service and service["instancename"] is not None and service["instancename"] != "":
+                command = "cf delete-service '" + service["instancename"] + "' -f"
+                message = "Delete CF service instance >" + service["instancename"] + "< from subaccount"
+                result = runShellCommand(btpUsecase, command, "INFO", message)
+        
+        log.info("Check deletion status for service instances")
+        # check status of deletion
+        search_every_x_seconds = btpUsecase.repeatstatusrequest
+        usecaseTimeout = btpUsecase.repeatstatustimeout
+        current_time = 0
+        allServicesDeleted = False
+        # Set the deletion status to "not deleted"
+        for service in accountMetadata["createdServiceInstances"]:
+            service["deletionStatus"] = "not deleted"
+        while usecaseTimeout > current_time and allServicesDeleted is False:
+            for service in accountMetadata["createdServiceInstances"]:
+                if "instancename" not in service:
+                    status = "deleted"
+                    service["deletionStatus"] = status
+                    log.info("no service instance available for service >" + service["name"] + "<. Deletion not needed.")
+                    continue
+                status = get_cf_service_deletion_status(btpUsecase, service)
+                if (status == "deleted"):
+                    log.success("service instance >" + service["instancename"] + "< for service >" + service["name"] + "< now deleted.")
+                    service["deletionStatus"] = "deleted"
+                else:
+                    service["deletionStatus"] = status
+            time.sleep(search_every_x_seconds)
+            current_time += search_every_x_seconds
+            allServicesDeleted = True
+            for service in accountMetadata["createdServiceInstances"]:
+                if service["deletionStatus"] != "deleted":
+                    allServicesDeleted = False
+        log.success("all service instances now deleted.")
 
     for environment in btpUsecase.definedEnvironments:
         if environment.name == "cloudfoundry":
